+++ main.tf
@@ -12,12 +12,49 @@ provider "aws" {
 ##########################
 resource "aws_vpc_peering_connection" "this" {
   provider      = "aws.this"
-  count         = "${(var.create_peering * (1 + var.cross_region_peering)) == "1" ? 1 : 0}"
+  count         = "${var.create_peering == "1" ? 1 : 0}"
   peer_owner_id = "${var.peer_account_id == "" ? data.aws_caller_identity.current.account_id : var.peer_account_id}"
   peer_vpc_id   = "${var.peer_vpc_id}"
   vpc_id        = "${var.this_vpc_id}"
-  auto_accept   = "${var.auto_accept_peering}"
+  peer_region   = "${var.peer_region == "" ? data.aws_region.current.name : var.peer_region}"
   tags          = "${var.tags}"
+}
+
+######################################
+# VPC peering accepter configuration #
+######################################
+resource "aws_vpc_peering_connection_accepter" "peer_accepter" {
+  provider                  = "aws.peer"
+  count                     = "${var.create_peering ? 1 : 0}"
+  vpc_peering_connection_id = "${aws_vpc_peering_connection.this.id}"
+  auto_accept               = "${var.auto_accept_peering}"
+  tags                      = "${merge(var.tags, map("Side", "Accepter"))}"
+}
+
+#######################
+# VPC peering options #
+#######################
+resource "aws_vpc_peering_connection_options" "this" {
+  provider                  = "aws.this"
+  vpc_peering_connection_id = "${aws_vpc_peering_connection_accepter.peer_accepter.id}"
+
+  requester {
+    allow_remote_vpc_dns_resolution  = "${var.this_dns_resolution}"
+    allow_classic_link_to_remote_vpc = "${var.this_link_to_peer_classic}"
+    allow_vpc_to_remote_classic_link = "${var.this_link_to_local_classic}"
+  }
+}
+
+resource "aws_vpc_peering_connection_options" "accepter" {
+  provider = "aws.peer"
+
+  vpc_peering_connection_id = "${aws_vpc_peering_connection_accepter.peer_accepter.id}"
+
+  accepter {
+    allow_remote_vpc_dns_resolution  = "${var.peer_dns_resolution}"
+    allow_classic_link_to_remote_vpc = "${var.peer_link_to_peer_classic}"
+    allow_vpc_to_remote_classic_link = "${var.peer_link_to_local_classic}"
+  }
 }
 
 ###################
@@ -25,7 +62,7 @@ resource "aws_vpc_peering_connection" "this" {
 ###################
 resource "aws_route" "this_routes_region" {
   provider                  = "aws.this"
-  count                     = "${(var.create_peering * (1 + var.cross_region_peering)) == 1 ? length(data.aws_route_tables.peer_vpc_rts.ids) : 0}"
+  count                     = "${var.create_peering == 1 ? length(data.aws_route_tables.peer_vpc_rts.ids) : 0}"
   route_table_id            = "${data.aws_route_tables.this_vpc_rts.ids[count.index]}"
   destination_cidr_block    = "${data.aws_vpc.peer_vpc.cidr_block}"
   vpc_peering_connection_id = "${var.peering_id == "" ? element(concat(aws_vpc_peering_connection.this.*.id, list("")), 0) : var.peering_id}"
@@ -36,53 +73,8 @@ resource "aws_route" "this_routes_region" {
 ###################
 resource "aws_route" "peer_routes_region" {
   provider                  = "aws.peer"
-  count                     = "${(var.create_peering * (1 + var.cross_region_peering)) == 1 ? length(data.aws_route_tables.peer_vpc_rts.ids) : 0}"
+  count                     = "${var.create_peering == 1 ? length(data.aws_route_tables.peer_vpc_rts.ids) : 0}"
   route_table_id            = "${data.aws_route_tables.peer_vpc_rts.ids[count.index]}"
   destination_cidr_block    = "${data.aws_vpc.this_vpc.cidr_block}"
   vpc_peering_connection_id = "${var.peering_id == "" ? element(concat(aws_vpc_peering_connection.this.*.id, list("")), 0) : var.peering_id}"
-}
-
-############################
-# VPC cross-region peering #
-############################
-resource "aws_vpc_peering_connection" "this_cross_region" {
-  provider      = "aws.this"
-  count         = "${(var.create_peering * var.cross_region_peering) == "1" ? 1 : 0}"
-  peer_owner_id = "${var.peer_account_id == "" ? data.aws_caller_identity.current.account_id : var.peer_account_id}"
-  peer_vpc_id   = "${var.peer_vpc_id}"
-  vpc_id        = "${var.this_vpc_id}"
-  peer_region   = "${var.peer_region}"
-}
-
-#####################################
-# Accepter's side of the connection #
-#####################################
-resource "aws_vpc_peering_connection_accepter" "peer_accepter" {
-  provider                  = "aws.peer"
-  count                     = "${(var.create_peering * var.cross_region_peering) == "1" ? 1 : 0}"
-  vpc_peering_connection_id = "${aws_vpc_peering_connection.this_cross_region.id}"
-  auto_accept               = true
-  tags                      = "${merge(var.tags, map("Side", "Accepter"))}"
-}
-
-###################
-# This Cross Region VPC Routes #
-###################
-resource "aws_route" "this_routes_cross_region" {
-  provider                  = "aws.this"
-  count                     = "${(var.create_peering * var.cross_region_peering) == "1" ? length(data.aws_route_tables.peer_vpc_rts.ids) : 0}"
-  route_table_id            = "${data.aws_route_tables.this_vpc_rts.ids[count.index]}"
-  destination_cidr_block    = "${data.aws_vpc.peer_vpc.cidr_block}"
-  vpc_peering_connection_id = "${var.peering_id == "" ? element(concat(aws_vpc_peering_connection.this_cross_region.*.id, list("")), 0) : var.peering_id}"
-}
-
-###################
-# Peer Cross Region VPC Routes #
-###################
-resource "aws_route" "peer_routes_cross_region" {
-  provider                  = "aws.peer"
-  count                     = "${(var.create_peering * var.cross_region_peering) == "1" ? length(data.aws_route_tables.peer_vpc_rts.ids) : 0}"
-  route_table_id            = "${data.aws_route_tables.peer_vpc_rts.ids[count.index]}"
-  destination_cidr_block    = "${data.aws_vpc.this_vpc.cidr_block}"
-  vpc_peering_connection_id = "${var.peering_id == "" ? element(concat(aws_vpc_peering_connection.this_cross_region.*.id, list("")), 0) : var.peering_id}"
 }